Import data transaksi

+++ domain/src/main/java/me/skripsi/domain/ui_models/MappingExtension.kt
@@ -14,7 +14,7 @@ fun DataTransaksi.toUiDataTransaksi() : UiDataTransaksi{
             namaBarang = namaBarang,
             golongan = golongan,
             stok = stok,
-            isDiskon = isDiskon,
+            isDiskon = isDiskon.toShort(),
             penjualan = penjualan,
             pembelian = pembelian
         )
